requirements.txt
python-libsbml>=5.20.5
sbmlutils>=0.9.0
sbmlpbkutils @ git+https://github.com/jwkruisselbrink/sbml-pbk-utils.git@v0.32.0
tellurium>=2.2.11.1
antimony>=3.1.0
seaborn==0.13.2
pandas>=2.3.1
parameterized>=0.9.0
nbconvert>=7.16.6
mkdocs>=1.6.1
mkdocs-material>=9.6.20
mkdocs-macros-plugin>=1.3.9
Jinja2>=3.1.6
PyYAML>=6.0
